pdf_api/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
reportlab>=3.6
django-environ>=0.10
